Minor code cleanup to get rid of unused members and methods
Classes wbase::ChannelShared (used to be wbase::SendChannelShared)
and wbase::TransmitData had an obsolete mechanism for setting up
result schema in the response messages to be sent to Czar. The current
implementation no longer uses that API. The obsilete API was removed
to avoid confusions as it wasn't obvious where the actual schema
settings were made.