Refactored and extended worker configuration service
The service is now available to the client code via shared pointer
that can be stored and used by classes as needed. The change reduces
the number of parameters which are sent around the code. In the new
version of the code only the shared pointer to the service is put
to where the configuration parameters are consumed.

Also added new parameters to support the file-based result delivery,
including: a location of the results folder, the number of BOOST ASIO
threads to run the QHTTP server at worker, and a selector for the desired
results delivery protocol.