Modified Czar to allow pulling result files from workers
The new version of the result merger at Czar would dynamically determine
which protocol should be used for pulling results from workers: SSI stream,
a file read via the XROOTD file protocol, or a file read via the HTTP
protocol. In case of the last two options, the merger would also also
utomatically tell a worker to delete the result file upon completion of
the merge (including the unsuccessful ones).

A choice of the delivery method is based on the optional
fileds in the Protobuf messages received from workers.
It's up to the workers to decide on what method to select.