Extended the Controller/Qserv workers status monitoring protocol
It's now possible to explicitly request info on tasks and restrict
a scope of the request to specific query identifiers in the status
requests sent by the Controller to workers.

Extended the Controller's REST API to allow passing the very same
optional parameters in the requests. Incremented th everson number
of the REST API to 19.