More reliable implementation for cancelling XROOTD/SSI request
The new implementation is based on the smart pointer to self
be stored by the Qserv request classes. It would guarantee the life
span of the request objects while request processing is still going
on. Otherwse, there is a chance of running into crashes. The request
cancellaton method has been added to replace SSI's Finished(true)
to do a proper clean up of the stored pointer.