Refactored configuration service of Czar
The service is now implemented as a singleton which is initialized
and loaded with the configuration parameters at a time when Czar
is being created. The rest of the Czar fetches values of the parameters
from the singleton. The refactoring aims at reducing the number of parameters
sent around in the code and to improve readability of the code.