Enforce large result limit at workers when fetching result sets
Note that the enforcement only applies to the file-based result
delivery protocol as this mechanism makes no sense when results
are streamed to Czar over the SSI protocol.

The large result conditions are reported to Czar with the code
util::ErrorCode::WORKER_RESULT_TOO_LARGE which would help Czar
to identify this specific condition and properly report the one
to a user.