Extended error reporting in the REST services
Added an option to send back to a client a warning along when
version mismatch is detected. The warnings are meant to be used
for reporting problems with parameter validations. One of the use
cases is when a client is attempting to use parameters introduced
in the future version of the API while reporting the older version.