Modified one test data set to test support for non-unique director keys
Added a new director table DeepSourceDupKeys, config files and the data.
A collection of rows to be ingested into the table intentionally has
duplicated rows as needed to test the new feature. The config files
were prepared with the same purpose.

Note that no actual queries would be launched against the table. The current
change is only meant to test the new ingest capability.