Capture MySQL thread identifiers for worker queries
The identifiers are captured for the worker task monitoring
purposes. They are used by the Web Dashboard for associating task
activities with the corresponding MySQL queries issued by the tasks.
The task monitoring page of the Dashboard has been extended to display
a value of the identifier (if it was set for the task).