Refined states of the worker tasks
Introduced an additional state STARTED to differentiate between
two events - when a task was started by a scheduler from the actual
start time of the corresponding MySQL query. The change is meant
to improve the monitoring of the worker tasks.
The corresponding monitoring page of the Web Dashboard was modified
as well.