Improved and extended implementation of the Czar monitoring
Extend the Histogram class to report the histogram label (id).
Added a histogram to capture the performance of the result file
reads at Czar. Added counters for the number of on-going
result file reads (file-based result delivery protocol only)
and merges. Fixed database merge histogram load.
Added integral counters for metrics reset at the start up time
of Czar. Added the timestamp (milliseconds) attribute to the JSON
object that returns a state of the monitoring parameters. The timestamps
are going to be used for the performance monitoring of Qserv.