Move class replica::AsyncTimer into the utility module
Refactored dependencies accordingly. The move allows re-using the
class by other Qserv modules w/o adding a direct depedency onto
the module 'replica' which was the original location of the class.

Fixed a bug in the implementation of the timer to allow subsequent
restarts. Unblocking implementation of the callback calling mechanism.
Extended a model of the callbacks to allow ordering automatic restarts
of the timer via the return value (boolean type) of the callbacks.
Minor refactpring of the implementation.