Temporary fix to the connection management of the query manager
The fix was introduced after discovering that the shared connection
usage approach in the original design of this and related classes
is highly unreliable. A connection shared by the queries seems
to be sensitive to the service startup ordering and occasional
disconnects. As a result of that the shared connection could just
stop working or it could not be alive from the very start of Czar.
It's also possible that the shared connection object is not properly
synchronized to allow using it in the multi-threaded environment
(which would be a bug). The last theory will be further investigated later.