Fixed a bug in the result folder cleanup at the worker start up time
In the original implementation of the file-based result delivery protocol,
workers were attempting to clean up unclaimed files left in workers'
result folders regardless of the protocol option. This clearly was
a mistake for the SSI protocol option where the folder wasn't
required to exist. As a result of this, the application posts
confusing warnings in the logging stream. This was fixed.

Another problem that was fixed was related to the protocol options HTTP
and XROOT where the application wouldn't abort right away if the required
folder didn't exist during the folder cleanup attempt. Not having this
folder available in this scenario means a configuration error or
a problem with the infrastructure.