Fixed a bug in the state transition logic of the worker tasks
In the previous code, the task were declared as "finished" after
being "booted" from a scheduler. In reality, the tasks could be
still in the active state processing the corresponding queries
and consuming resources (memory, CPU, MySQL connections).
This change made the task monitoring correct and allowed
to uncover tasks that were previously seen as "invisible".

The startup sequence of the worker tasks has been modified in the same
way to declare the task processing at a point where the actual
processing begins.