Added "relaxed" memory management mode at Qserv worker
The memory management node is a variation of MemManNone except
it won't report a failure to (memory) lock tables where such
locking is required. With this mode, the scheduler would not
enforce additional restrictions on the number of queries that are
processed in parallel. In case of using the non-relaxed manager
MemManNone the processing concurrency is forced to be 1 per queue
regardles of what's specified in the worker's configuration or
the number of the hardware threads on a machine.