Fixed a bug in the result reading algorithm at Czar.
In the previous version, the reading of files over HTTP is terminated
when a TCP connection is closed by a worker. However, the latter might
not happen if the HTTP server is configured to reuse sockets.
In the new version of the code the file transfer is finished when
the reader gets enough rows (as specified in the worker response
summary message).

Also, did some minor refactoring in the row counting code.