Web Dashboard: extended Czar monitoring displays to allow many Czars
NOTE: The only problem that has not been addressed by the commit is
a navigation form the active queries display to the query progression
(burn down) plot. The navigation requires czar names to be reported
by the Replication Controller along with the active query info.
This is hard to do due to a known problem with Czar instances having
the "dual identity", which includes both some original (arbitrary name)
injected to Czar by a frontend (mysql-proxy or http-based one) and
an automatically generated numeric identifier for that input name.
The number identifiers are used elsewhere withing Czar implementation.
And the names of Czars are used by the Qserv monitoring and management
systems. Teh problem will be addressed in a separate effort