Extended query completion/cancellation protocol to send Czar id
This extension prevents interference at workers when they are
processing query management requests sent by different Czars
in the multi-Czar deployments. The naming convention for
the worker-side partial result files has been extended as well
to include the numeric identifier of a Czar responsible
for the corresponding query.
Added the unit test for the file name parser