Improved initialization of the HTTP connection pool based on Czar config
Fixed incorrect comments. Minor reinforcement in the merging handler
to insure the corresponding job is still alive while mering the next batch
of the partial result rows into the destination table. Minor improvementes
in the parameters of methods reporting the number of rows in a result set.
In the previous version of the code the limit was set to 4 bilion rows which may
be potentially exceeded for the LSST-scale catalogs.