Optimized resultset extraction algorithm in Qserv worker
The new algorithm significantly increases parallelism
in result handling of the sub-chunk-based queries to allow
each task of a composite (multi-task) job to extract result sets
simultaneously. The file-based recoding stream (wbase::FileChannleShared)
is now locked for a very short duration of time needed to write
a serialized Protobuf message into the output file of a job.

Another improvement of this change is regarding the overall design
of the code. Responsibilities of classes wdb::QueryRunner and
wbase::FileChannelShared have been streamlined in the new version
of the code. Specifically, all operations with MySQL are now done
by wdb::QueryRunner and they are not spread between the result
streaming class wbase::FileChannelShared and the former as it was
done in the previous code.