Extended Qserv Cli and dedicated entry point for testing HTTP frontend
NOTE: This work is yet to be complete. The current implementation adds
the fair amount of code duplication, which may or may not be a problem,
depending on the long-term plans for eliminating the MySQL frontend.
A substantial refactoring effort may be needed if the MySQL proxy-based
frontend stayed for a long period of time. Should such reafactoring
decision be made it would rather be done in a context of a separate ticket.