Streamlined and enchanced database API for transactions and contributions
Eliminated duplicate methods.

Extended the filtering model for transactions and contributions in the API.
It's now possible to specify the optional filters for the transaction and
contribution states. Added the optional limit for the total number of
contributions pulled from the database. The new filter is used by
the Web Dashboard to prevent overloading the UI and to avoid triggering
heavy database queries.