Fixed formatting of the C++ files
The change was made by running the latest version of clang-format
in the build container. Apparently, the tool pulled into the latest
version of the container has different rules for the C++ code.
The older verson of the tool was:
"clang-format version 17.0.6 (AlmaLinux OS Foundation 17.0.6-5.el9)"
The new version is now:
"clang-format version 18.1.8 (AlmaLinux OS Foundation 18.1.8-3.el9)"