Minor code clean up and refactoring in the result merging code
Minor refactoring of class RowBuffer to remove code that was not relevant
in a context of the class. It was probaly created as a result of prior
experimentations. The current design and the implementation of the class
has nothing to do with Protobuf types.

Moved the MySQL-specific string "escape" functions into a more appropriate
header. Created a dedicated unit test for the functions.

The last change was made to allow shifting CSV generation from Czar to workers.
The older code live in the Czar-specific module which shouldn't be referenced
from the worker's modules.