Minor clean up of unused identifier in the worker code
The identity mechanism was introduced in the original implementation of
the worker-side "shared channels" which was used for streaming partial
results of queries from workers Czars over the XROOTD/SSI streaming protocol.
The identifiers were used for a purpose of finding bugs and race conditions
in the worker code. This is no longer needed since the data are now moved
to Czar(s) over the HTTP protocol.