Refactored and extended the CSV buffer classes of the mysql module
Changes were made to make the names of classes and variables be consistent with
their actual roles. The content of some header files were also reorganized
to move unimportant private definitions into the CC file.

Added a specialized class for forwarding a CSV-formated buffer (of bytes).
The new class is needed for ingesting the worker-generatd CSV payload
into the result tables.

Added a thread-safe stream class for passing the CSV-formatted records
from the HTTP file reader to the MySQL table loader.

Added the unit tests for the new classes.