Refactored and extended management threads launched by Czar
Put the garbage collection algorithm into a separate module. Refactored
the implementation of the algorithm to always run in a loop within
a dedicated persistent thread.

Implemented a separate garbage collection algorithm for result sets of
the completed (failed or aborted) ASYNC queries.

Extended configuration service of Czar with a parameter that defines
an age of the ASYNC queries to be evaluated by the garbage collection
algorithm.