Possible fix for XROOTD issues #2479 #2481
ATTENTION: This is a temporary commit that is relying directly
on the PR in the xrootd Git package. The PR will disappear after
being merged into the master branch of the package. A permanent solution
will be to upgrade Qserv to depend on XROOTD v5.9 when it will be
available.