Refactored the query progress classes and dependencies
The first goal of the changes was to name classes and variables according to
their actual roles. The second goal was to avoid ambiguity in the
naming convention betwen the query status enum qmeta::QInfo::Qstatus and former
class qmeta::QStatus.

The chaos in naming the relevant variables has been also eliminated. The
names have been streamlined to improve the overall readability of the code.