Reinforced active query monitoring REST service
The new version of the service adds a filter to select queries
that are still executing. This eliminates false-positive entries
from being reported by the service when entries in the query progress
table are no longer active. The change is also ment to eliminate
tose entries from being displayed in the Qserv Dashboard.