Garbage collect aborted queries from the progress table
The changes are about the Czar metadata table QProgress. When Czar
frontend restarts the table would still keep obsolete rows of
the queries that are normally aborted during at the restart time from
the table QInfo. This commit completes the operation to put the progress
table into the clean state after each restart. Note that the operation
only affects queries submitted via the corresponding Czar frontend.

Added a management thread to periodically clean up the obsolete
entries from the table QProgress. This is needed because the current
implementation of Czar has a logic flow allowing some queries to stay in
the table. This results in confusing reports made by the Qserv
Monitoring System.