Extended REST services for ingesting data and creating tables
The services now allow to specify the desired charset and/or collation
names when creating tables in Qserv or ingesting user tables in Qserv.
The protocol between the Replication Controller and the Replication
workers has been extended as well.
The relevant request, job and applications of the Replication Framework
have been extended as well.