DM-52370: Extended replica allocation services
The new services extend chunk allocations (placement) to include all
existing replicas for the chunks in question instead of just one
allocation per chunk. The extended services are meant to be used by
the ingest workflows to push multiple replicas of a chunk in scenarios
where the replication level is bigger than 1 and multiple replicas
already exist in Qserv. The older services would fail in such scenarios.

Minor refactoring on the older code.

Fixed a bug in the JSON schema of  of the chunks allocation service
that was returning the key "location" instead of "locations". Migrated
the integration test accordingly.

Fixed a bug in the integration test.

Migrated the documentation on the core Ingest API.