Reinforced implementation of the Replication system
The fix is meant to prevent crashes of the Replicaton Controller
and the workers after deleting databases. In the previous
implementation Controller's request and jobs, as well as the worker
requests might be still attempting to process outstanding requests
related to the previously deleted databases. The problem exists
due to the distributed architecture of Qserv.